Support of Docker over SSH
